+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Features
+* **@dlr-eoc/services-map-state**
+ - Allow to use `projjson` in `IProjDef`.
+
 # [16.0.0](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v16.0.0) (2026-05-29) (update dependencies, new library ngx-ukis-ui-clarity and ngx-ukis-utilities, replace core-ui)
 
 ### Breaking Changes